fix(worker): decouple live previews from gateway acknowledgements
