+++ src/worker/embedded-agent-live.runtime.ts
@@ -104,45 +104,6 @@ function boundLiveEvent(event: WorkerLiveEvent): WorkerLiveEvent {
   return bounded;
 }
 
-function coalescePendingLiveEvent(pending: WorkerLiveEvent[], event: WorkerLiveEvent): boolean {
-  const index = pending.length - 1;
-  const previous = pending[index];
-  if (!previous) {
-    return false;
-  }
-  if (previous.kind === "assistant" && event.kind === "assistant") {
-    pending[index] = boundLiveEvent({
-      kind: "assistant",
-      payload: { ...event.payload, delta: event.payload.text, replace: true },
-    });
-    return true;
-  }
-  if (previous.kind === "thinking" && event.kind === "thinking") {
-    if (event.payload.text === "" && event.payload.delta === "") {
-      return false;
-    }
-    pending[index] = boundLiveEvent({
-      kind: "thinking",
-      payload: {
-        text: event.payload.text,
-        delta: `${previous.payload.delta}${event.payload.delta}`,
-      },
-    });
-    return true;
-  }
-  if (
-    previous.kind === "tool" &&
-    previous.payload.phase === "update" &&
-    event.kind === "tool" &&
-    event.payload.phase === "update" &&
-    previous.payload.toolCallId === event.payload.toolCallId
-  ) {
-    pending[index] = boundLiveEvent(event);
-    return true;
-  }
-  return false;
-}
-
 function readAssistantText(message: AgentMessage): string {
   if (message.role !== "assistant") {
     return "";
@@ -170,57 +131,22 @@ type WorkerLiveClient = {
 type WorkerLiveRuntime = {
   handleSessionEvent: (event: AgentSessionEvent) => void;
   enqueueRunFailure: (failure: { aborted: boolean; error: Error }) => void;
-  flush: () => Promise<void>;
   emitTerminal: () => Promise<void>;
 };
 
 export function createWorkerLiveRuntime(client: WorkerLiveClient): WorkerLiveRuntime {
-  const pendingLiveEvents: WorkerLiveEvent[] = [];
-  let liveDrain: Promise<void> | undefined;
   let liveDegraded = false;
-  const startLiveDrain = () => {
-    if (liveDrain || liveDegraded || pendingLiveEvents.length === 0) {
-      return;
-    }
-    liveDrain = (async () => {
-      while (true) {
-        const event = pendingLiveEvents.shift();
-        if (!event) {
-          return;
-        }
-        await client.emit(event);
-      }
-    })()
-      .catch(() => {
-        // Live events are preview-only; transcript commits and inference stay authoritative.
-        liveDegraded = true;
-        pendingLiveEvents.length = 0;
-      })
-      .finally(() => {
-        liveDrain = undefined;
-        startLiveDrain();
-      });
-  };
   const enqueueLive = (event: WorkerLiveEvent) => {
     if (liveDegraded) {
       return;
     }
     try {
-      const bounded = boundLiveEvent(event);
-      if (!coalescePendingLiveEvent(pendingLiveEvents, bounded)) {
-        pendingLiveEvents.push(bounded);
-      }
-      startLiveDrain();
+      void client.emit(boundLiveEvent(event)).catch(() => {
+        // Preview loss must not block inference, transcript durability, or finishing.
+        liveDegraded = true;
+      });
     } catch {
       liveDegraded = true;
-      pendingLiveEvents.length = 0;
-    }
-  };
-  const flush = async () => {
-    let drain = liveDrain;
-    while (drain) {
-      await drain;
-      drain = liveDrain;
     }
   };
   const startedAt = Date.now();
@@ -358,5 +284,5 @@ export function createWorkerLiveRuntime(client: WorkerLiveClient): WorkerLiveRun
     }
     await client.emit(boundLiveEvent(terminalLiveEvent));
   };
-  return { handleSessionEvent, enqueueRunFailure, flush, emitTerminal };
+  return { handleSessionEvent, enqueueRunFailure, emitTerminal };
 }
